OEP-2025: Fixed the inherent brokenness in the nested record architecture (and a memory leak).
At one point, way long ago, the nested record architecture was working properly. When we last
merged with SC official, it somehow got broked.

This commit essentially restores the original functionality AND fixes a memory leak (that existed
in the original functionality).

AS OF RIGHT NOW, all of the unit tests in the datastore framework are PASSING.

DON'T COMMIT NEW CODE THAT BREAKS THEM!!!

Unless you immediately commit a fix, of course.