OEP-2025: Fixes broken nested record architecture.
This reintroduces reverted commits and further modifies the functionality of
the SC datastore framework (specifically related to nested records/stores).

It fixes the inherent broken nature of the nested record architecture, which
had been the cause of memory leaks, data corruption and all manner of odd
behavior in the datastore layer.

This is a large commit with many changes.

Unit tests have been added to cover new functionality, and existing unit tests
have been updated to reflect the changes.

All tests are currently passing.