OEP-9616: Fixed bug in creation of parent/childRecords mappings for nested stores.
They need to be deeply cloned (which isn't a big deal because they're small) in
order to avoid erroneous propagation of nested store changes to the parent.

We could have also handled this lazily in the _lock() method, but again, the
structures are small, so it really doesn't hurt to just do a full deep clone.